refactor mutations out of schema.js into schema files
