+++ server.js
@@ -31,7 +31,7 @@ const checkJwt = jwt({
 
 app.use(
   '/backend',
-  //checkJwt,
+  checkJwt,
   graphqlHTTP({
     schema,
     graphiql: true,
